Rename directory name: `src/models` -> `src/model` #196
