+++ src/model/util.rs
